--- a/tests/Benchmarks/System.Text.Primitives/AsciiCasing.cs
+++ b/tests/Benchmarks/System.Text.Primitives/AsciiCasing.cs
@@ -1,34 +1,37 @@
-﻿using BenchmarkDotNet.Attributes;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using BenchmarkDotNet.Attributes;
 using System.Buffers;
 
 namespace System.Text.Primitives.Benchmarks
 {
     public class AsciiCasing
     {
-        private byte[] bytes;
-        private byte[] output;
+        private byte[] _bytes;
+        private byte[] _output;
 
         [Params("/plaintext", "text/plain,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7")]
         public string Text;
 
-
         [GlobalSetup]
         public void Setup()
         {
-            bytes = Encoding.ASCII.GetBytes(Text);
-            output = new byte[bytes.Length];
+            _bytes = Encoding.ASCII.GetBytes(Text);
+            _output = new byte[_bytes.Length];
         }
 
         [Benchmark]
-        public OperationStatus AsciiToLowerInPlace() => Buffers.Text.Encodings.Ascii.ToLowerInPlace(bytes, out int bytesChanged);
+        public OperationStatus AsciiToLowerInPlace() => Buffers.Text.Encodings.Ascii.ToLowerInPlace(_bytes, out int bytesChanged);
 
         [Benchmark]
-        public OperationStatus AsciiToLower() => Buffers.Text.Encodings.Ascii.ToLower(bytes, output, out int bytesChanged);
+        public OperationStatus AsciiToLower() => Buffers.Text.Encodings.Ascii.ToLower(_bytes, _output, out int bytesChanged);
 
         [Benchmark]
-        public OperationStatus AsciiToUpperInPlace() => Buffers.Text.Encodings.Ascii.ToUpperInPlace(bytes, out int bytesChanged);
+        public OperationStatus AsciiToUpperInPlace() => Buffers.Text.Encodings.Ascii.ToUpperInPlace(_bytes, out int bytesChanged);
 
         [Benchmark]
-        public OperationStatus AsciiToUpper() => Buffers.Text.Encodings.Ascii.ToUpper(bytes, output, out int bytesChanged);
+        public OperationStatus AsciiToUpper() => Buffers.Text.Encodings.Ascii.ToUpper(_bytes, _output, out int bytesChanged);
     }
 }