Below are results of the changes suggested in #1738. The first run has the change for Ascii.ToLowerInPlace() which is why you'll notice a discrepancy.
```
// * Summary *

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17134.191 (1803/April2018Update/Redstone4)
Intel Core i7-7820HQ CPU 2.90GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
Frequency=2835933 Hz, Resolution=352.6176 ns, Timer=TSC
.NET Core SDK=2.1.400
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT


              Method |                 Text |      Mean |     Error |    StdDev |
-------------------- |--------------------- |----------:|----------:|----------:|
 AsciiToLowerInPlace |           /plaintext |  21.76 ns | 0.3919 ns | 0.3474 ns |
        AsciiToLower |           /plaintext |  29.70 ns | 0.2521 ns | 0.2358 ns |
 AsciiToUpperInPlace |           /plaintext |  28.22 ns | 0.1585 ns | 0.1238 ns |
        AsciiToUpper |           /plaintext |  30.00 ns | 0.3067 ns | 0.2394 ns |
 AsciiToLowerInPlace | text/(...)q=0.7 [86] | 112.71 ns | 2.2612 ns | 2.7770 ns |
        AsciiToLower | text/(...)q=0.7 [86] | 179.32 ns | 3.0066 ns | 2.6653 ns |
 AsciiToUpperInPlace | text/(...)q=0.7 [86] | 184.81 ns | 3.7301 ns | 6.1287 ns |
        AsciiToUpper | text/(...)q=0.7 [86] | 184.16 ns | 3.8889 ns | 5.7002 ns |

// * Summary *

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17134.191 (1803/April2018Update/Redstone4)
Intel Core i7-7820HQ CPU 2.90GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
Frequency=2835933 Hz, Resolution=352.6176 ns, Timer=TSC
.NET Core SDK=2.1.400
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT


              Method |                 Text |      Mean |     Error |    StdDev |
-------------------- |--------------------- |----------:|----------:|----------:|
 AsciiToLowerInPlace |           /plaintext |  21.62 ns | 0.3449 ns | 0.3226 ns |
        AsciiToLower |           /plaintext |  20.93 ns | 0.2101 ns | 0.1965 ns |
 AsciiToUpperInPlace |           /plaintext |  21.48 ns | 0.3310 ns | 0.2934 ns |
        AsciiToUpper |           /plaintext |  21.28 ns | 0.2685 ns | 0.2380 ns |
 AsciiToLowerInPlace | text/(...)q=0.7 [86] | 110.34 ns | 1.8941 ns | 1.6790 ns |
        AsciiToLower | text/(...)q=0.7 [86] |  95.50 ns | 2.6976 ns | 2.5233 ns |
 AsciiToUpperInPlace | text/(...)q=0.7 [86] | 109.41 ns | 0.9965 ns | 0.8834 ns |
        AsciiToUpper | text/(...)q=0.7 [86] |  90.93 ns | 0.3209 ns | 0.3002 ns |

```