@kassemsandarusi 2018-08-19
Build seemed to fail, but it seems unrelated to my changeset.

I believe 
```
            if (!_disposed)
                _timer.Change(PollingInterval, Timeout.Infinite);
```
https://github.com/dotnet/corefxlab/blob/6f5861cfd5914d889086177c79ea1b4df240d774/src/System.IO.FileSystem.Watcher.Polling/System/IO/PollingFileSystemWatcher.cs#L189-L208

Failed with the following:
```
02:00:47 Starting test execution, please wait...
02:00:54 The active test run was aborted. Reason: Unhandled Exception: System.ObjectDisposedException: Cannot access a disposed object.
02:00:54    at System.Threading.TimerQueueTimer.Change(UInt32 dueTime, UInt32 period)
02:00:54    at System.IO.PollingFileSystemWatcher.TimerHandler(Object context) in D:\j\workspace\windows_nt_re---4abe806d\src\System.IO.FileSystem.Watcher.Polling\System\IO\PollingFileSystemWatcher.cs:line 208
02:00:54    at System.Threading.ExecutionContext.RunInternal(ExecutionContext executionContext, ContextCallback callback, Object state)
02:00:54 --- End of stack trace from previous location where exception was thrown ---
02:00:54    at System.Threading.TimerQueueTimer.CallCallback()
02:00:54    at System.Threading.TimerQueueTimer.Fire()
02:00:54    at System.Threading.TimerQueue.FireNextTimers()
```