@ahsonkhan 2018-08-20
> Build seemed to fail, but it seems unrelated to my changeset.

Known issue: https://github.com/dotnet/corefxlab/issues/420
cc @JeremyKuhne, @MisinformedDNA