@ahsonkhan 2018-08-20
@dotnet-bot test Innerloop Windows_NT Release Build and Test

Maybe we should disable this test for the time being.